trying again

+++ src/ladder.cpp
@@ -8,7 +8,7 @@
 #include <fstream>
 using namespace std;
 
-typedef unsigned long long ull;
+typedef uint64_t ull;
 
 
 
@@ -71,6 +71,7 @@ bool edit_distance_within(const std::string& str1, const std::string& str2, int 
 bool is_adjacent(const string& word1, const string& word2)
 {
     return edit_distance_within(word1, word2, 1);
+
 }
 // uses bfs
 //creates a queue of stacks
